ocr/requirements.txt

python-dotenv==1.0.1
imageio==2.37.0
numpy==2.2.2
onnxruntime==1.20.1
openai==1.60.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
pillow==11.1.0
python-Levenshtein==0.26.1
PyYAML==6.0.2
videodb==0.2.10
grpcio==1.60.1
mistralai==1.5.1
google-genai==1.4.0
